rename package names to follow standard java naming conventions
