--- src/test/java/tests/MobileAppTests/Objects/LinkAccountsPageObjects.java
@@ -1,13 +1,0 @@
-package tests.MobileAppTests.Objects;
-
-import io.appium.java_client.MobileElement;
-import io.appium.java_client.pagefactory.AndroidFindBy;
-
-public class LinkAccountsPageObjects {
-    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/tv_patient_name")
-    public MobileElement patientName;
-
-    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/btn_search")
-    public MobileElement linkSelected;
-
-}
