--- src/test/java/tests/MobileAppTests/Pages/ConsentArtifactPage.java
@@ -1,24 +1,0 @@
-package tests.MobileAppTests.Pages;
-
-import tests.MobileAppTests.Objects.ConsentArtifactPageObjects;
-import tests.MobileAppTests.Utils.WaitUtils;
-import io.appium.java_client.AppiumDriver;
-import io.appium.java_client.pagefactory.AppiumFieldDecorator;
-import org.openqa.selenium.support.PageFactory;
-
-public class ConsentArtifactPage extends WaitUtils {
-
-    ConsentArtifactPageObjects consentArtifactPageObjects;
-
-    public ConsentArtifactPage(AppiumDriver driver) {
-        super(driver);
-        consentArtifactPageObjects = new ConsentArtifactPageObjects();
-        PageFactory.initElements(new AppiumFieldDecorator(driver), consentArtifactPageObjects);
-        waitForElement(consentArtifactPageObjects.grantButton);
-    }
-
-    public OTPPage grantConsent() {
-        consentArtifactPageObjects.grantButton.click();
-        return new OTPPage(driver);
-    }
-}
