--- src/test/java/tests/MobileAppTests/Utils/Gestures.java
@@ -1,31 +1,0 @@
-package tests.MobileAppTests.Utils;
-
-import io.appium.java_client.AppiumDriver;
-import io.appium.java_client.MobileElement;
-import org.openqa.selenium.Point;
-import org.openqa.selenium.interactions.PointerInput;
-import org.openqa.selenium.interactions.Sequence;
-
-import java.time.Duration;
-import java.util.Arrays;
-
-public class Gestures {
-
-    public void verticalSwipe(AppiumDriver driver, MobileElement element) {
-        Point source = element.getCenter();
-        PointerInput finger = new PointerInput(PointerInput.Kind.TOUCH, "finger");
-
-        Sequence swipe = new Sequence(finger, 0);
-
-        swipe.addAction(finger.createPointerMove(Duration.ofMillis(0),
-                PointerInput.Origin.viewport(),
-                source.x / 2, source.y));
-        swipe.addAction(finger.createPointerDown(PointerInput.MouseButton.LEFT.asArg()));
-        swipe.addAction(finger.createPointerMove(Duration.ofMillis(30),
-                PointerInput.Origin.viewport(),
-                source.x / 2, source.y - 800));//Move the pointer up
-
-        swipe.addAction(finger.createPointerUp(PointerInput.MouseButton.LEFT.asArg()));
-        driver.perform(Arrays.asList(swipe));
-    }
-}
