+++ src/test/java/tests/mobiletests/pageobjects/BasePageObjects.java
@@ -1,0 +1,13 @@
+package tests.mobiletests.pageobjects;
+
+import io.appium.java_client.MobileElement;
+import io.appium.java_client.pagefactory.AndroidFindBy;
+
+public class BasePageObjects {
+
+    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/btn_next")
+    public MobileElement nextButton;
+
+    @AndroidFindBy(id = "in.ndhm.phr.debug:id/btn_skip")
+    public MobileElement skipButton;
+}
