+++ src/test/java/tests/mobiletests/pageobjects/SearchLinkProviderObjects.java
@@ -1,0 +1,19 @@
+package tests.mobiletests.pageobjects;
+
+import io.appium.java_client.MobileElement;
+import io.appium.java_client.pagefactory.AndroidFindBy;
+
+public class SearchLinkProviderObjects {
+
+    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/sv_provider")
+    public MobileElement providerInput;
+    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/provider_name")
+    public MobileElement providerName;
+    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/tv_mobile_number")
+    public MobileElement verifiedMobileNo;
+    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/btn_search")
+    public MobileElement confirmProvider;
+    @AndroidFindBy(id = "in.projecteka.jataayu.debug:id/lbl_mobile_number")
+    public MobileElement allRecordsLinkedText;
+
+}
