+++ src/test/java/tests/mobiletests/tests/LoginTest.java
@@ -1,0 +1,20 @@
+package tests.mobiletests.tests;
+
+import tests.apitests.helpers.PropertiesCache;
+import tests.mobiletests.pages.LoginPage;
+import tests.mobiletests.utils.BaseDriver;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+public class LoginTest extends BaseDriver {
+
+    String userName = PropertiesCache.getInstance().getProperty("mobileUser");
+    String password = PropertiesCache.getInstance().getProperty("mobilePassword");
+
+    @Test
+    public void loginUser() {
+
+        boolean status = new LoginPage(driver).loginUser(userName, password).validateHomePage();
+        Assert.assertTrue(status);
+    }
+}
