+++ src/test/java/tests/mobiletests/utils/IRunner.java
@@ -1,0 +1,9 @@
+package tests.mobiletests.utils;
+
+import tests.mobiletests.pageobjects.BasePageObjects;
+import tests.mobiletests.pageobjects.LoginPageObjects;
+
+public interface IRunner {
+    void moveSplashScreens(BasePageObjects basePageObjects);
+    void login(LoginPageObjects loginPageObjects, String username, String password);
+}
