+++ src/test/java/tests/mobiletests/utils/Patient.java
@@ -1,0 +1,12 @@
+package tests.mobiletests.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Patient {
+    public String patientName;
+    public String mobile;
+    public String gender;
+}
